Split symlink methods out of FileSystemUtilities
* Refocused existing symlink utility
* Added more bootstrap --hoist snapshots
