+++ test/FileSystemUtilities.js
@@ -5,8 +5,6 @@ const path = require("path");
 // mocked modules
 const fs = require("fs-extra");
 const pathExists = require("path-exists");
-const cmdShim = require("cmd-shim");
-const readCmdShim = require("read-cmd-shim");
 const ChildProcessUtilities = require("../src/ChildProcessUtilities");
 
 // helpers
@@ -17,11 +15,7 @@ const FileSystemUtilities = require("../src/FileSystemUtilities");
 
 jest.mock("fs-extra");
 jest.mock("path-exists");
-jest.mock("cmd-shim");
-jest.mock("read-cmd-shim");
 jest.mock("../src/ChildProcessUtilities");
-
-const linkRelative = (from, to) => path.relative(path.dirname(to), from);
 
 describe("FileSystemUtilities", () => {
   describe(".mkdirp()", () => {
@@ -177,176 +171,5 @@ describe("FileSystemUtilities", () => {
       FileSystemUtilities.unlinkSync(filePath);
       expect(fs.unlinkSync).lastCalledWith(filePath);
     });
-  });
-
-  describe(".isSymlink()", () => {
-    if (process.platform !== "win32") {
-      it("returns false when filePath is not a symlink", () => {
-        const filePath = path.resolve("./not/a/symlink");
-
-        fs.lstatSync.mockReturnValueOnce({
-          isSymbolicLink: () => false,
-        });
-
-        expect(FileSystemUtilities.isSymlink(filePath)).toBe(false);
-      });
-
-      it("returns resolved path of an existing symlink", () => {
-        const original = path.resolve("./packages/package-2");
-        const filePath = path.resolve("./packages/package-1/node_modules/package-2");
-
-        fs.lstatSync.mockReturnValueOnce({
-          isSymbolicLink: () => true,
-        });
-        fs.readlinkSync.mockReturnValueOnce(linkRelative(original, filePath));
-
-        expect(FileSystemUtilities.isSymlink(filePath)).toBe(original);
-      });
-    } else {
-      it("returns false when filePath is not a symlink (windows)", () => {
-        const filePath = path.resolve("./not/a/symlink");
-
-        fs.lstatSync.mockReturnValueOnce({
-          isSymbolicLink: () => false,
-          isFile: () => false,
-        });
-
-        expect(FileSystemUtilities.isSymlink(filePath)).toBe(false);
-      });
-
-      it("returns resolved path of an existing symlink (windows)", () => {
-        const original = path.resolve("./packages/package-2");
-        const filePath = path.resolve("./packages/package-1/node_modules/package-2");
-
-        fs.lstatSync.mockReturnValueOnce({
-          isSymbolicLink: () => true,
-          isFile: () => false,
-        });
-        fs.readlinkSync.mockReturnValueOnce(linkRelative(original, filePath));
-
-        expect(FileSystemUtilities.isSymlink(filePath)).toBe(original);
-      });
-
-      it("returns false when filePath is not a shimmed executable", () => {
-        const filePath = path.resolve("./packages/package-1/node_modules/.bin/package-2");
-
-        fs.lstatSync.mockReturnValueOnce({
-          isSymbolicLink: () => false,
-          isFile: () => true,
-        });
-        readCmdShim.sync.mockImplementationOnce(() => {
-          throw new Error("ENOTASHIM");
-        });
-
-        expect(FileSystemUtilities.isSymlink(filePath)).toBe(false);
-      });
-
-      it("returns resolved path of a shimmed executable", () => {
-        const original = path.resolve("./packages/package-2/cli.js");
-        const filePath = path.resolve("./packages/package-1/node_modules/.bin/package-2.cmd");
-
-        fs.lstatSync.mockReturnValueOnce({
-          isSymbolicLink: () => false,
-          isFile: () => true,
-        });
-        readCmdShim.sync.mockReturnValueOnce(linkRelative(original, filePath));
-
-        expect(FileSystemUtilities.isSymlink(filePath)).toBe(original);
-      });
-    }
-  });
-
-  describe(".symlink()", () => {
-    fs.lstat.mockImplementation(callsBack("ENOENT"));
-    fs.unlink.mockImplementation(callsBack());
-    fs.symlink.mockImplementation(callsBack());
-
-    if (process.platform !== "win32") {
-      it("creates relative symlink to a directory", done => {
-        const src = path.resolve("./packages/package-2");
-        const dst = path.resolve("./packages/package-1/node_modules/package-2");
-        const type = "junction"; // even in posix environments :P
-
-        FileSystemUtilities.symlink(src, dst, type, () => {
-          try {
-            expect(fs.unlink).not.toBeCalled();
-            expect(fs.symlink).lastCalledWith(linkRelative(src, dst), dst, type, expect.any(Function));
-            done();
-          } catch (ex) {
-            done.fail(ex);
-          }
-        });
-      });
-
-      it("creates relative symlink to an executable file", done => {
-        const src = path.resolve("./packages/package-2/cli.js");
-        const dst = path.resolve("./packages/package-1/node_modules/.bin/package-2");
-        const type = "exec";
-
-        FileSystemUtilities.symlink(src, dst, type, () => {
-          try {
-            expect(fs.unlink).not.toBeCalled();
-            expect(fs.symlink).lastCalledWith(linkRelative(src, dst), dst, "file", expect.any(Function));
-            done();
-          } catch (ex) {
-            done.fail(ex);
-          }
-        });
-      });
-
-      it("overwrites an existing symlink", done => {
-        const src = path.resolve("./packages/package-2");
-        const dst = path.resolve("./packages/package-1/node_modules/package-2");
-        const type = "junction"; // even in posix environments :P
-
-        fs.lstat.mockImplementationOnce(callsBack()); // something _does_ exist at destination
-
-        FileSystemUtilities.symlink(src, dst, type, () => {
-          try {
-            expect(fs.unlink).lastCalledWith(dst, expect.any(Function));
-            expect(fs.symlink).lastCalledWith(linkRelative(src, dst), dst, type, expect.any(Function));
-            done();
-          } catch (ex) {
-            done.fail(ex);
-          }
-        });
-      });
-    } else {
-      it("creates command shim to an executable file", done => {
-        const src = path.resolve("./packages/package-2/cli.js");
-        const dst = path.resolve("./packages/package-1/node_modules/.bin/package-2");
-        const type = "exec";
-
-        cmdShim.mockImplementationOnce(callsBack());
-
-        FileSystemUtilities.symlink(src, dst, type, () => {
-          try {
-            expect(fs.lstat).not.toBeCalled();
-            expect(cmdShim).lastCalledWith(src, dst, expect.any(Function));
-            done();
-          } catch (ex) {
-            done.fail(ex);
-          }
-        });
-      });
-
-      it("always uses absolute paths when creating symlinks", done => {
-        const src = path.resolve("./packages/package-2");
-        const dst = path.resolve("./packages/package-1/node_modules/package-2");
-        const type = "junction"; // only _actually_ matters in windows
-
-        fs.lstat.mockImplementationOnce(callsBack()); // something _does_ exist at destination
-
-        FileSystemUtilities.symlink(src, dst, type, () => {
-          try {
-            expect(fs.unlink).lastCalledWith(dst, expect.any(Function));
-            expect(fs.symlink).lastCalledWith(src, dst, type, expect.any(Function));
-            done();
-          } catch (ex) {
-            done.fail(ex);
-          }
-        });
-      });
-    }
   });
 });
