+++ test/LinkCommand.js
@@ -1,7 +1,7 @@
 "use strict";
 
 // mocked or stubbed modules
-const FileSystemUtilities = require("../src/FileSystemUtilities");
+const createSymlink = require("../src/utils/create-symlink");
 
 // helpers
 const callsBack = require("./helpers/callsBack");
@@ -11,22 +11,19 @@ const normalizeRelativeDir = require("./helpers/normalizeRelativeDir");
 // file under test
 const lernaLink = require("./helpers/command-runner")(require("../src/commands/LinkCommand"));
 
+jest.mock("../src/utils/create-symlink");
+
 // object snapshots have sorted keys
 const symlinkedDirectories = testDir =>
-  FileSystemUtilities.symlink.mock.calls.map(args => ({
-    _src: normalizeRelativeDir(testDir, args[0]),
-    dest: normalizeRelativeDir(testDir, args[1]),
-    type: args[2],
+  createSymlink.mock.calls.map(([src, dest, type]) => ({
+    _src: normalizeRelativeDir(testDir, src),
+    dest: normalizeRelativeDir(testDir, dest),
+    type,
   }));
 
 describe("LinkCommand", () => {
-  const fsSymlink = FileSystemUtilities.symlink;
-  beforeEach(() => {
-    FileSystemUtilities.symlink = jest.fn(callsBack());
-  });
-  afterEach(() => {
-    FileSystemUtilities.symlink = fsSymlink;
-  });
+  // the underlying implementation of symlinkDependencies
+  createSymlink.mockImplementation(callsBack());
 
   describe("with local package dependencies", () => {
     it("should symlink all packages", async () => {
